refactor: KEEP-356 move template-autocomplete-filter under lib/workflow/editor
Follows up on the merge resolution. The variable-suggestion filter
helpers added to lib/ root in PR #944 are editor-domain code and
belong with the other editor helpers under lib/workflow/editor/.
